Unify diffusion visualizer rendering on a single source of truth
Move the DiffusionGemma4 rendering primitives — display-width-exact word
wrapping and the flash-free in-place _CanvasRedrawer with throttling and
alternate-screen tail clipping — into models/diffusion_visualizer.py, and
rebuild DiffusionUnmaskingVisualizer (nemotron, llada2) on top of the
shared redrawer instead of its own clear-then-reprint logic. The gemma4
visualizer keeps only its model-specific text-stream composition and
engine-handler patch, importing the primitives from the shared module
(re-exported under the same names for existing importers).

Net behavior change for nemotron/llada2: wrapping is now display-width
aware (CJK-safe), redraws are flash-free, and oversized alternate-screen
canvases clip to the tail instead of scrolling. Gemma4 behavior is
unchanged; _CanvasRedrawer.draw() gains an optional force flag that
defaults off.
